+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Technical - Upgrade ESLint rules.
+- Technical - Ensure that all files follow the ESLint rules.
 
 ## RELEASE 5.4.0 - 2019-11-29
 ### Added